+++ src/scandeval/constants.py
@@ -4,7 +4,7 @@
 DUMMY_FILL_VALUE = 100
 
 
-GENERATIVE_MODEL_TASKS = ["text-generation", "conversational", "text2text-generation"]
+GENERATIVE_MODEL_TASKS = ["text-generation", "text2text-generation"]
 
 
 GENERATIVE_DATASET_TASKS = [